FEAT: Merge sovereign sub round end files into a single file
